Wrap systemd calls with exception handles.
This fixes the issue with ooniprobe trying to start the agent when it
has updated to a much more recent version (>= 2.1.0) and then the
updater scripts run.

This affects all lepidopter devices that are running ooniprobe 1.4.1
(lepidopter version v0.2.1-alpha) and were booted for the first time
after the release of ooniprobe 2.1.0 (2016-12-07).